+++ components/server/src/model/session.py
@@ -1,6 +1,6 @@
 """Session model class."""
 
-from datetime import datetime
+from datetime import datetime, timezone
 from typing import Union, cast
 
 
@@ -12,8 +12,10 @@ class Session:
 
     def is_valid(self) -> bool:
         """Return whether the session is valid."""
-        expiration_datetime = cast(datetime, self.__session_data.get("session_expiration_datetime", datetime.min))
-        return bool(expiration_datetime > datetime.now())
+        expiration_datetime = cast(
+            datetime, self.__session_data.get("session_expiration_datetime", datetime.min)
+        ).replace(tzinfo=timezone.utc)
+        return bool(expiration_datetime > datetime.now(timezone.utc))
 
     def is_authorized(self, authorized_users: list[str]) -> bool:
         """Return whether the session's user is an authorized user."""